reorganize tests and remove duplicates sources

+++ testDrm/src/Linux/Bindings/EvdevClass.cs
@@ -25,7 +25,7 @@
 // OUT OF OR IN CONNECTION WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS IN
 // THE SOFTWARE.
 //
-using OpenTK;
+
 using Linux;
 using Crow;
 
@@ -416,7 +416,7 @@ namespace Linux.oldEvDev
             unsafe
             {
                 sbyte* pname = stackalloc sbyte[129];
-                int ret = Libc.ioctl(fd, EvdevIoctl.Name128, new IntPtr(pname));
+				int ret = Libc.ioctl(fd, (uint)EvdevIoctl.Name128, new IntPtr(pname));
                 name = new string(pname);
                 return ret;
             }
@@ -429,7 +429,7 @@ namespace Linux.oldEvDev
             {
                 fixed (EvdevInputId* pid = &id)
                 {
-                    return Libc.ioctl(fd, EvdevIoctl.Id, new IntPtr(pid));
+					return Libc.ioctl(fd, (uint)EvdevIoctl.Id, new IntPtr(pid));
                 }
             }
         }
